'''Attention for all users''' '''This Update work only if you have a Nemesis svn >= 092!!!! ''' '''If you have a SVN version < 092, or SSL < 84, follow this:''' * Flash SecondStage Loader 84 * Reboot * Flash Image SVN 098r6 * Reboot * Update to new release on-line '''If you have a SVN version >= 092, or SSL = 84''' * You can flash directly a new 098r6 * Update to new release on-line
'''Release 100r8'''
* Update radio.mvi screen logo
* Add enable/disable deepstandby function on front power button long press
Blue->Yellow->Setup Console/RC/Enigma->Enable deepstandby on front power button long press
* Add new choice "do nothing" on RC 'Action on long powerbutton press'
* Add Unwanted Reboot Manager
Blue->Yellow->Setup Console/RC/Enigma->Go to standby after (unwanted) reboot
* Update skin default
* Replace HD-Glass-15 skin to old skin default
* Fix HDD info on Device Status Panel
* Update Plugin StayUP, Fix load epg on Mode 1
* Fix Mode 2 on dm7020hd now enigma run in both mode 1 & 2
* Load EPG in a cache without restart Enigma2 on E2 3.2.1 ('''Thank's Dr.Best''')
No plugin epgloader need!!!! Plese do not install it!!!!
If you have epgloader plugin in your box, please remove it, before upgrade
* Update Linux Kernel (Small fix)
* Update bootlogo
* Update [http://git.opendreambox.org/?p=ghost/opendreambox-1.6;a=commit;h=e34658b20a500d492f9201fff2d3b1d5493e7f78 dreambox-dvb-modules] to 2011109 dm500hd, 800, 800se, 8000
* Update [http://git.opendreambox.org/?p=ghost/opendreambox-1.6;a=commit;h=505ea246d4b4757c79528e81b771ffd71a6bd59d dreambox-dvb-modules] to 2011121 dm7020hd
* Update Mode 1 enigma2 to 3.2.1
* Partially rewrited Green Panel (Epg Control Center)
* Fix crash on Mode 2 on System Info (Yellow and Blue Button)
* Fix crash on Mode 2 using new setting created by Blindscan (I hope :))
* Fix loading kernel module for rt3070 based WIFI Adapter
* Fix skin to support new Harddisk Manager
* Update Italian translation
* Update enigma2-plugins
After updating, and reboot, you do mount HDD with new Harddisk Manager!
* Menu->Setup->System->Harddisk

Download full image [http://nemesis.tv/NemesisImage/ here].


git-svn-id: svn://192.168.1.30/Nemesis/trunk@100 70059d43-2406-4b27-96de-136f9efb2096

+++ create_e2_package.sh
@@ -60,8 +60,8 @@ create ()
 
 	echo -n "Prepare package..."
 	cd $nemesis_e2_pack
-	mv usr/lib/enigma2/python/enigma.py usr/lib/enigma2/python/enigma.py.nemesis
-	mv usr/bin/enigma2 usr/bin/enigma2.nemesis
+	[ -e usr/lib/enigma2/python/enigma.py ] && mv usr/lib/enigma2/python/enigma.py usr/lib/enigma2/python/enigma.py.nemesis
+	[ -e usr/bin/enigma2 ] && mv usr/bin/enigma2 usr/bin/enigma2.nemesis
 
 	cd $srcpatchedpath/enigma2_$MACHINE
 	mv usr/lib/enigma2/python/enigma.py  $nemesis_e2_pack/usr/lib/enigma2/python/enigma.py.32
@@ -75,6 +75,7 @@ create ()
 	rm -rf enigma2_$MACHINE/usr/lib/enigma2/python/*.py*
 	rm -rf enigma2_$MACHINE/usr/lib/enigma2/python/Components
 	rm -rf enigma2_$MACHINE/usr/lib/enigma2/python/Screens
+	rm -rf enigma2_$MACHINE/usr/lib/enigma2/python/Tools
 	rm -rf enigma2_$MACHINE/usr/lib/enigma2/python/Plugins/Extensions/GraphMultiEPG
 	rm -rf enigma2_$MACHINE/usr/lib/enigma2/python/Plugins/SystemPlugins/SoftwareManager
 	rm -rf enigma2_$MACHINE/usr/lib/enigma2/python/Plugins/SystemPlugins/WirelessLan
@@ -85,6 +86,7 @@ create ()
 	cp -f $nemesis_e2_pack/usr/lib/enigma2/python/*.py* enigma2_$MACHINE/usr/lib/enigma2/python/
 	cp -rf $nemesis_e2_pack/usr/lib/enigma2/python/Components enigma2_$MACHINE/usr/lib/enigma2/python/
 	cp -rf $nemesis_e2_pack/usr/lib/enigma2/python/Screens enigma2_$MACHINE/usr/lib/enigma2/python/
+	cp -rf $nemesis_e2_pack/usr/lib/enigma2/python/Tools enigma2_$MACHINE/usr/lib/enigma2/python/
 	cp -rf $nemesis_e2_pack/usr/lib/enigma2/python/Nemesis enigma2_$MACHINE/usr/lib/enigma2/python/
 	cp -rf $nemesis_e2_pack/usr/lib/enigma2/python/Plugins/Extensions/GraphMultiEPG enigma2_$MACHINE/usr/lib/enigma2/python/Plugins/Extensions/
 	cp -rf $nemesis_e2_pack/usr/lib/enigma2/python/Plugins/SystemPlugins/SoftwareManager enigma2_$MACHINE/usr/lib/enigma2/python/Plugins/SystemPlugins/
